% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetrize.R
\name{asymmetrise}
\alias{asymmetrise}
\alias{asymmetrize}
\title{Add all missing comparisons between two columns}
\usage{
asymmetrise(.data, .x, .y)

asymmetrize(.data, .x, .y)
}
\arguments{
\item{.data}{a tidy \code{data.frame} or \code{tibble}}

\item{.x, .y}{the data to add all comparisons between (ie. will be the
x and y-axes for \code{geom_asymmat}}
}
\value{
a data table with new rows for the added comparisons
}
\description{
This function prepares input data for \code{geom_asymmat} by
    adding in any missing comparisons to be plotted.
}
\examples{

df <- data.frame(a = c("A", "B"),
                 b = c("C", "A"),
                 untouched = c(1, 2))
df

asymmetrise(df, a, b)

}
