% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asymmetric_tibble.R
\name{col_swap}
\alias{col_swap}
\title{Copy a tibble, but flip .x and .y}
\usage{
col_swap(.data, .x, .y)
}
\arguments{
\item{.data}{A tibble}

\item{.x, .y}{Column names to swap}
}
\value{
The same tibble with the values in columns \code{.x} and \code{.y} swapped
}
\description{
Copy a tibble, but flip .x and .y
}
\examples{
suppressMessages(library(dplyr))
suppressMessages(library(magrittr))

tib <- tibble(col_a = c(1:5),
              col_b = LETTERS[1:5],
              not_changed = runif(5))

tib

switched_ab <- tib \%>\%
    col_swap(col_a, col_b)

switched_ab

}
