% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_continuous_asym.R
\name{continuous_scale_asym}
\alias{continuous_scale_asym}
\title{Continuous scale constructor for ggasym}
\usage{
continuous_scale_asym(aesthetics, scale_name, palette, na.value, guide,
  ...)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with}

\item{scale_name}{The name of the scale}

\item{palette}{A palette function that when called with a numeric vector
with values between 0 and 1 returns the corresponding values in the
range the scale maps to.}

\item{na.value}{Missing values will be replaced with this value.}

\item{guide}{A function used to create a guide or its name. See guides()
for more info.}

\item{...}{other input is passed on to \code{ggplot2::continuous_scale()};
see \code{?ggplot2::continuous_scale} for complete documentation}
}
\value{
a scales object \code{FINISH THIS}
}
\description{
This is a this wrapper around \code{continuous_scale()} from
    the ggplot2 package.
}
