% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{factor_is_greater}
\alias{factor_is_greater}
\title{Determines if the level of a is greater than that of b}
\usage{
factor_is_greater(a, b)
}
\arguments{
\item{a, b}{Two same-length, same-leveled vectors of type \code{factor}}
}
\value{
a single boolean vector of type \code{logical}
}
\description{
Determines if the level of a is greater than that of b
}
\examples{
first <- c("J", "O", "S", "H")
last <- c("C", "O", "O", "K")
first <- factor(first, LETTERS)
last <- factor(last, LETTERS)

factor_is_greater(first, last)

}
