% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_asymmat.R
\name{geom_asymat}
\alias{geom_asymat}
\title{Generate an asymmetric matrix with different fill values for top-left
    and bottom-right triangles}
\usage{
geom_asymat(.data, x, y, fill_tl, fill_br)
}
\description{
Generate an asymmetric matrix with different fill values for top-left
    and bottom-right triangles
}
