% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_asymmat.R
\name{geom_asymmat}
\alias{geom_asymmat}
\title{Asymmetrically filled symmetric matrix (using ggplot)}
\usage{
geom_asymmat(.data, x, y, fill_tl, fill_br, color_tl = NA,
  color_br = NA, lab_tl = NULL, lab_br = NULL)
}
\arguments{
\item{.data}{data table}

\item{x, y}{x and y-axes column names}

\item{fill_tl, fill_br}{column names for top-left and bottom-right fill}

\item{color_tl, color_br}{colors for the lines; use \code{NA} for no color}

\item{lab_tl, lab_br}{labels for the scales; use \code{NULL} (default) to use
the column name (ie. \code{fill_tl} or \code{fill_br}) or empty strings
(\code{""}) for no label}
}
\value{
a ggplot object of an asymmetrically-colored \code{x} x \code{y}
    matrix with \code{fill_tl} data coloring the top-left triangle and
    \code{fill_br} coloring the bottom-left triangle.
}
\description{
Generate an asymmetric matrix with different fill values for top-left
    and bottom-right triangles as a \code{ggplot} object
}
