% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_asymmat.R
\docType{data}
\name{GeomAsymmat}
\alias{GeomAsymmat}
\title{GeomAsymmat}
\format{An object of class \code{GeomAsymmat} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 7.}
\usage{
GeomAsymmat
}
\description{
A ggproto object for the ggasym package and used by \code{geom_asymmat}
}
\section{Warning}{

This \code{GeomAsymmat} is very much still in development and warrant to change
    without notice. Use at your own risk. If dependent on \code{GeomAsymmat}
    it is advisable to include tests with a cached version to test for
    equivalence.
}

\keyword{datasets}
