% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetrize.R
\name{get_other_combs}
\alias{get_other_combs}
\title{Get all combinations of values between two vectors}
\usage{
get_other_combs(x, y)
}
\arguments{
\item{x, y}{two vectors}
}
\value{
data.frame of other posibble combinations stored in \code{Var1} and
    \code{Var2} for \code{x} and \code{y}, respectively
}
\description{
Get all combinations of the values in vectors x and y that
    aren't already there.
}
\examples{
get_other_combs(LETTERS[1:2], LETTERS[1:2])

}
