% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetrise.R
\name{is_grouped}
\alias{is_grouped}
\title{Is a data table grouped?}
\usage{
is_grouped(.data)
}
\arguments{
\item{.data}{input \code{data.frame} or \code{tibble}}
}
\value{
boolean
}
\description{
Determines if the input data frame or tibble is grouped
    (using \code{dplyr::group_by()}
}
\examples{
df <- data.frame(x = c(1:5), g = c(1,1,2,2,2))
is_grouped(df)

is_grouped(dplyr::group_by(df, g))

}
