% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moexpr_data}
\alias{moexpr_data}
\alias{mouse_diet}
\title{Diet of mice for gene expression data}
\format{\code{moexpr_data} is a tibble (18,596 x 49)
\describe{
    \item{gene}{gene name for the row}
    \item{GSM994790 ... GSM994837}{expression data for each mouse}
}
\code{mouse_diet} is a tibble (48 x 2)
\describe{
    \item{mouse}{the ID of the mouse}
    \item{diet}{the level of fat in the diet}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/sites/GDSbrowser/}
}
\usage{
moexpr_data

mouse_diet
}
\description{
\code{mouse_diet} contains the mouse ID and diet information for the mice
    used in this study.
}
\details{
\code{moexpr_data} contains the gene expression of each mouse (one mouse
    per column)

This data was obtained from the Gene Expression Omnibus (GEO) under DataSet
    Record GDS6248.

"Analysis of livers of C57BL/6J mice fed a high fat diet for up to 24 weeks.
    Significant body weight gain was observed after 4 weeks. Results provide
    insight into the effect of high fat diets on metabolism in the liver."
}
\examples{
moexpr_data
mouse_diet

}
\keyword{datasets}
