% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asymmetric_tibble.R
\name{prepare_asymmetic_tibble}
\alias{prepare_asymmetic_tibble}
\title{Returns a tibble of the diagonal of a hypothetical symmetrix matrix}
\usage{
prepare_asymmetic_tibble(.data, .x, .y, .tl, .br)
}
\arguments{
\item{.data}{a tibble}

\item{.x, .y}{the \emph{x} and \emph{y} column names of the hypothetical
matrix}

\item{...}{column names to group by}
}
\value{
a tibble where everything is the same except that the two selected
    columns are set equal
}
\description{
Returns a tibble where everything is the same except that the two selected
columns are set equal. It is used to set diagonal values
}
\examples{
"NOT USED"
Complete the \\math{n x n} matrix and make asymmetric

}
