% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_level}
\alias{which_level}
\title{Determine the level of a value in a vector of type \code{factor}}
\usage{
which_level(x)
}
\arguments{
\item{x}{vectors of type \code{factor}}
}
\value{
a vector holding the corresponding level of the input factors
}
\description{
Determine the level of a value in a vector of type \code{factor}
}
\examples{
first <- factor(c("J", "O", "S", "H"), LETTERS)
which_level(first)

}
