% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asymmetric_tibble.R
\name{col_swap}
\alias{col_swap}
\title{Copy a tibble, but flip .x and .y}
\usage{
col_swap(.data, .x, .y)
}
\arguments{
\item{.data}{A tibble}

\item{.x, .y}{Column names to swap}
}
\value{
The same tibble with the values in columns \code{.x} and \code{.y} swapped
}
\description{
Copy a tibble, but flip .x and .y
}
\examples{
library(dplyr)
library(magrittr)

tib <- tibble(col_a = c(1:5),
              col_b = LETTERS[1:5],
              not_changed = runif(5))

tib

#> # A tibble: 5 x 3
#>   col_a col_b not_changed
#>   <int> <chr>       <dbl>
#> 1     1 A           0.904
#> 2     2 B           0.603
#> 3     3 C           0.632
#> 4     4 D           0.937
#> 5     5 E           0.850

switched_ab <- tib \%>\%
    col_swap(col_a, col_b)

switched_ab

#> # A tibble: 5 x 3
#>   col_a col_b not_changed
#>   <chr> <int>       <dbl>
#> 1 A         1       0.904
#> 2 B         2       0.603
#> 3 C         3       0.632
#> 4 D         4       0.937
#> 5 E         5       0.850

}
