% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_asymmat.R
\name{geom_asymat}
\alias{geom_asymat}
\title{Generate an asymmetric matrix with different fill values for top-left
    and bottom-right triangles}
\usage{
geom_asymat(.data, x, y, fill_tl, fill_br)
}
\arguments{
\item{.data}{data table}

\item{x, y}{x and y-axes column names}

\item{fill_tl, fill_br}{column names for topleft and bottom-right fill}
}
\value{
a ggplot object of an asymmetrically-colored \code{x} x \code{y}
    matix with \code{fill_tl} data coloring the top-left triangle and
    \code{fill_br} coloring the bottom-left triangle.
}
\description{
TODO: add description
}
