% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggasym.R
\docType{package}
\name{ggasym}
\alias{ggasym}
\alias{ggasym-package}
\title{ggasym: Asymmetric Matrix Plotting in ggplot}
\description{
This package plots a symmetric matrix with two different fill
aesthetics for the top-left and bottom-right triangles and a third along the diagonal.
It operates within the Grammar of Graphics paradigm implemented in ggplot.
}
