% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_asymmetric_tibble.R
\name{prepare_asymmetric_tibble}
\alias{prepare_asymmetric_tibble}
\title{Complete the \eqn{n \times n} matrix and make asymmetric}
\usage{
prepare_asymmetric_tibble(.data, .x, .y, .tl, .br)
}
\arguments{
\item{.data}{a tibble with at least four columns for the x and y-axes
(\code{.x} and \code{.y}) and the top-left and bottom-right values
(\code{.tl} and \code{.br})}

\item{.x, .y}{x and y-axes column names}

\item{.tl, .br}{top-left and bottom right column names}
}
\description{
This function is designed to make the "other half" of the \eqn{n \times n}
matrix. It copies the matrix, reversing \code{.x} and \code{.y}, followed by
only keeping the top-left (\code{.tl}) and bottom-right (\code{br}) values
in the correct rows.
}
\examples{
library(tibble)

tib <- tibble(a = LETTERS[1:5],
              b = LETTERS[10:14],
              val_1 = c(0.70, 0.05, 0.14, 0.60, 0.83),
              val_2 = c(0.78, 0.25, 0.74, 0.26, 0.16))

tib

prepare_asymmetric_tibble(tib, a, b, val_1, val_2)

}
