% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_gradient.R
\name{scale_fill_br_gradient}
\alias{scale_fill_br_gradient}
\title{Gradient colour scales for bottom-right triangle in geom_asymmat}
\usage{
scale_fill_br_gradient(..., low = "#132B43", high = "#56B1F7",
  space = "Lab", na.value = "grey50", guide = "colourbar",
  aesthetics = "fill_br")
}
\arguments{
\item{...}{arguments passed on to \code{continuous_scale_asym}}

\item{low, high}{the colors to represent low and high values}

\item{space}{colour space in which to calculate gradient. Must be "Lab" -
other values are deprecated.}

\item{na.value}{color of missing (\code{NA}) values}

\item{guide}{Type of legend. Use "colourbar" for continuous colour bar, or
"legend" for discrete colour legend.}

\item{aesthetics}{Character string or vector of character strings listing
the name(s) of the aesthetic(s) that this scale works with. For now,
leave the default alone, though I plan to reinstate the standard ggplot
system here, eventually.}
}
\description{
This dictates a gradient color scheme for the bottom-right
    (\code{br}) of a \code{geom_asymmat} ggplot geom.
    \code{scale\_\*\_tl\_gradient} creates a two colour gradient (low-high),
    \code{scale\_\*\_tl\_gradient2} creates a diverging colour gradient
    (low-mid-high), \code{scale\_\*\_tl\_gradientn} creates a n-colour
    gradient.
}
\examples{
# vignette(ggasym)

}
