% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetrize.R
\name{swap_cols}
\alias{swap_cols}
\title{Swap columns in a data.frame}
\usage{
swap_cols(.data, .x, .y)
}
\arguments{
\item{.data}{a data.frame (or tibble) object}

\item{.x, .y}{column names to switch}
}
\value{
a data.frame (or tibble) object with \code{.x} and \code{.y} swapped
}
\description{
Swap columns \code{.x} and \code{.y} in \code{.data}.
}
\examples{
df <- data.frame(a = c("A", "B"),
                 b = c("C", "D"),
                 untouched = c(1, 2))
df

swap_cols(df, a, b)

}
